serverside_thesis/requirements.txt
Django==1.10
django-cors-headers==2.1.0
djangorestframework==3.6.2
ez-setup==0.9
olefile==0.44
Pillow==4.3.0
